Add eslint linter
Refers to issue: #28

Due to various pull requests over the years, coding style lucks a
specific standard. This commit address this issue by introducing
a linter to the project, specifically eslint. By providing a linter,
we are able to adhere to certain style guidelines, thus increasing
the maintainability.

Eslint rules are managed by the eslint-config-airbnb package.
Right now linting is supported only in the src/ directory. In the
future, this will likely need to be extend in spec/ folder as well.

eslint-config-airbnb package comes with the corresponding preset.
This preset highly suggest the use of ES6 features. However, those
rules are not taken into consideration to keep backward compatibility
with legacy systems that do not support ES6. In the future, this needs
to be addressed, either by disabling those rules or by dropping support
in legacy pre-ES6 environments.