Include spec/ folder to eslint
Refers to issue: #28

- Lint files under spec/ folder with eslint.
- Add ES6 syntax to test files under spec/ folder to pass
linter tests.

This commit limits the ability of CIs that are configured to
use node versions that do not support es6 basic features
(like arrow functions and const, let keyword). In those
enviroments the build will fail. In the future this has to
be addressed either by updating the node version in CIs, or
by switching back to ES5 code